+++ modules/budget/versions.tf
@@ -20,14 +20,14 @@ terraform {
   required_providers {
     google = {
       source  = "hashicorp/google"
-      version = "~> 4.28"
+      version = ">= 4.28, < 6"
     }
   }
 
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-project-factory:budget/v14.2.0"
+    module_name = "blueprints/terraform/terraform-google-project-factory:budget/v14.4.0"
   }
   provider_meta "google-beta" {
-    module_name = "blueprints/terraform/terraform-google-project-factory:budget/v14.2.0"
+    module_name = "blueprints/terraform/terraform-google-project-factory:budget/v14.4.0"
   }
 }