+++ modules/shared_vpc_access/versions.tf
@@ -19,17 +19,17 @@ terraform {
   required_providers {
     google = {
       source  = "hashicorp/google"
-      version = ">= 3.43, < 5.0"
+      version = ">= 3.43, < 6"
     }
     google-beta = {
       source  = "hashicorp/google-beta"
-      version = ">= 3.43, < 5.0"
+      version = ">= 3.43, < 6"
     }
   }
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-project-factory:shared_vpc_access/v14.2.0"
+    module_name = "blueprints/terraform/terraform-google-project-factory:shared_vpc_access/v14.4.0"
   }
   provider_meta "google-beta" {
-    module_name = "blueprints/terraform/terraform-google-project-factory:shared_vpc_access/v14.2.0"
+    module_name = "blueprints/terraform/terraform-google-project-factory:shared_vpc_access/v14.4.0"
   }
 }